Fixed handling of timouts and tokens
Currently there are CancellationTokenSource's created in the RestRequestImpl, which are not doing the job expected. HttpClient by default sets a timeout of 100 seconds, this is not overridden anywhere. What's happening is the tokens are being created in RestRequestImpl passed through to SendAsync, and once that 100 seconds is hit inside HttpClient, it is cancelling the token, which in turn cancels the linked token for the request and the whole thing fails. To fix this I have made it so that HttpUtil takes a timeout inside its constructor, which sets the timeout of the HttpClient. I changed all of the times to be TimeSpan classes for cleanliness and to reduce confusion between setting values in seconds vs ms.

Within HttpUtil.SendAsync there was also no handling for response being null, which was causing null exceptions on failure of a responce due to the token cancelling instead of being able to retry, so I added that.