Removed explicitly defined 'NETFRAMEWORK' preprocessor symbol from csproj; this fixes #308 and #536
NETFRAMEWORK is a standard preprocessor symbol that is already defined for builds that target .NET Framework (but not .NET Core builds); therefore defining this symbol in the csproj (with an attempt to exclude it for dotnet-core builds using a regex) is not necessary.

The regex filter was not working as expecting, causing this symbol to be defined for all builds including dotnet core builds.

See here for a list of standard preprocessor symbols:
https://learn.microsoft.com/en-us/dotnet/core/tutorials/libraries#how-to-multitarget