SNOW-777930 Bug fix: .NET application terminates when client_session_keep_alive=true due to unhandled exception (#642)
https://github.com/snowflakedb/snowflake-sdks-drivers-issues-teamwork/issues/307

There were mainly two issues:
1. unhandled exception from heartbeat session renew crashes upper
application
2. heartbeat queue growth possibly uncontrolled

The root cause of issue 1 is that when trying to send heart beat for a
session, there could be a bad timing that the session token expired so
renew session is triggered, and the session is closed right after so
renew session could fail, and the unhandled exception was thrown there.
Removing the exception throw will fix this issue.

There are multiple reason for issue 2 though.
- SFSession.Close() doesn't remove the session out of heart beat queue.
https://github.com/snowflakedb/snowflake-connector-net/blob/v2.0.22/Snowflake.Data/Core/SFSession.cs#L258
- SFSnowflakeConnection.Dispose() set disposed flag even when the
connection is pooled, so when the connection is reused by a new
connection and disposed again, nothing would happen as the disposed flag
is set to true already so the session was leaked.
https://github.com/snowflakedb/snowflake-connector-net/blob/v2.0.22/Snowflake.Data/Client/SnowflakeDbConnection.cs#L335
To fix issue 2, always call stopHeartBeatForThisSession() when close the
session, also in SFSnowflakeConnection.Dispose() only set disposed when
it's not pooled.

There are other bug fixes as well:
- CLIENT_SESSION_KEEP_ALIVE in connection string. It's actually not
working as the driver didn't send it to server as session parameter
- pooling timeout was reset each time when the connection is reused and
pooled again. This could make the connection never expire and
application could get error when the session expired. Set pooling
timeout only at the first time it's pooled.

Some other minor changes:
- add logging of session id.
- add some internal functions in HeartBeatBackground for test purpose.
- remove usage of codecov as it's deprecated and fail the test run.
- Ignore known test failure as it's been there for months and we have
[sdk issue
#210](https://github.com/snowflakedb/snowflake-sdks-drivers-issues-teamwork/issues/210)
for that already.