A command may be executed within an unexpected transaction.
If Close or Dispose is called for a Connection with an uncomplete transaction and this Connection is reused from the ConnectionPool, SQL commands will be executed in the old transaction scope.
Uncompleted transaction should be rolled back before Close. In addition, if there is any chance of a transaction surviving, it should not be added to the Pool.
In other cases, an explicit BEGIN may be unknowingly returned to the pool, so connections should be checked before they are reused.