Bug fix for 401 error with empty session token (#656)
https://github.com/snowflakedb/snowflake-sdks-drivers-issues-teamwork/issues/362

The root cause was that when OpenAsync() not finish the login process
the application close the connection already. In such case the driver
pool the invalid connection (session is empty) without any validation,
and when it get reused by other connection, the query request failed
with 401 error because the session token is empty.

Fix it by adding check of whether the session token is empty before
pooling the connection.
Also added logging for session id to get better tracking of the activity
for each session in case of using multiple connections in parallel.