Bug fix for "Session no longer exists" issue with connection pooling (#666)
For issue#416
The root cause is the improper design of connection pooling feature. It
was pooling SnowflakeDBConnection instances and making copy when it's
reused from pooling, so it's always possible that the connection hold by
different threads could point to the same underlying session.

In the case of issue 416 with `Session no longer exists` error, thread A
calls close() to close the connection and the connection get pooled,
thread B reuse the connection right after and get the connection out of
the pool. Then thread A calls Dispose() on the connection and since it's
not pooled, it close the session while it still being used by thread B,
and that's why thread B got error of `Session no longer exists`

The error was `Authentication token has expired` in earlier versions I
think that's because we made fix for other issues of connection pooling
which changed the behavior but the issue still there.

Pool the underlying session instead of the connection instance itself
should be able to solve all the timing issue and make the logic more
simple and clear. No copy on the session instance and a session should
be always either in pool or hold by one thread.

Test cases added for connection pooling with concurrent connections.