remove winhttphandler (#674)
Bug fix for issue#365

The issue was introduced by using WinHttpHandler as a replacement for
HttpClientHandler, which was found not a good replacement due to some
unexpected behavior changes.
.e.g the root cause of issue#365 likely is the same as issue#404, that
WinHttpHandler needs different time out settings.
The purpose of using WinHttpHandler is to fix an issue that when using
proxy and set INSECURE_MODE=false, the connection could fail with SSL
error.
Per discussion in the issue#365 that actually could be customer proxy
setting issue, also using WinHttpHandler doesn't really fix the issue as
it's been reported on Linux as well.

The decision here is to remove WinHttpHandler, to remove all unexpected
behavior changes with it. For the proxy issue if customer raise it again
we can check whether it's customer setting issue. In case we do need a
fix on driver side we could find a solution that fix in on all
platforms.