SNOW-835635 rollback session on connection pooled (#691)
This change introduces an automated rollback during connection close and
being pooled.
The rollback will be done on condition that a transaction got started
with its constructor or BeginTransaction method of a connection
(explicit transaction). If an error on such a rollback occurs the
connection won't get added back to the pool to ensure that pool does
never provide a dirty (with active transaction) connection.
In case of a transaction started with SnowflakeDbCommand the driver
won't keep track of it for performance reasons.