SNOW-902611 move control over creation of session to session pool (#783)
Moving a code responsible for establishing a new session to the session
pool class as a prerequisite step to introduction of a new connection
pool implementation.
Fixed some flaky tests which were unrelated to the pool but were failing
when mocked connection (with some intentional throw exception behavior)
was put back to the pool by another test.

- [x] Code compiles correctly
- [x] Code is formatted according to [Coding
Conventions](../CodingConventions.md)
- [ ] Created tests which fail without the change (if possible)
- [x] All tests passing (`dotnet test`)
- [x] Extended the README / documentation, if necessary
- [x] Provide JIRA issue id (if possible) or GitHub issue id in PR name